# Add fast decode plan for flashinfer mla